Add files via upload
This is the matlab ".m" file that is used to simulate the "PAM" method considering a VFC-enabled scenarios. The final "response latencies" are stored in variable "averages". However, note that we used the final "response latencies" are stored in variable "averages" for reference and based on these values, we filled a corresponding "excel" sheet with probable "response latencies" that can be shown by an "PAM" method.